+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `help` and `version` actions.
 - `-v` verbose option. Normal mode output is now cleaner and more focused.
 
 ### Changed
@@ -21,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Zsh 5.2 does not recognize the `:P` modifier. Replace it by `:A`.
 - Also compile and clean .zprofile among the startup files.
-- Don't fail on `zimfw clean-dumpfile` when there's nothing to remove.
+- Don't fail on `clean-dumpfile` when there's nothing to remove.
 
 ## [1.0.0] - 2020-01-07