+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `help` and `version` actions.
 - `-v` verbose option. Normal mode output is now cleaner and more focused.
-- Check the latest version using `git ls-remote`.
+- Asynchronously check the latest version every 30 days. This can be disabled
+  with `zstyle ':zim' disable-version-check yes`.
 
 ### Changed
 - When upgrading, download latest release asset instead of raw file from the